Move GritSparseMHA to components/grit_layers.py
Start a dedicated module for the GRIT graph-transformer layers; the
name avoids clashing with the GritTransformer model in models/gnn/grit.
components.layers re-exports the name so existing imports and saved
model configs keep working.
